# Speed up hashing for GridQubit, LineQubit, and NamedQubit