@maffoo 2023-11-17
Updated to optimized `NamedQubit` as well; first-time hashing sped up from 1.78us  to 553ns (3.2x speedup) and repeated hashing sped up from 220ns to 162ns (1.36x speedup).

On master:
```python
In [2]: nq = cirq.NamedQubit("abc")

In [3]: %timeit hash(nq)
220 ns ± 14.1 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)

In [4]: %timeit hash(cirq.NamedQubit("abc"))
1.78 µs ± 13.9 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)
```

On this branch:
```
In [2]: nq = cirq.NamedQubit("abc")

In [3]: %timeit hash(nq)
162 ns ± 4.14 ns per loop (mean ± std. dev. of 7 runs, 10,000,000 loops each)

In [4]: %timeit hash(cirq.NamedQubit("abc"))
553 ns ± 1.16 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)
```